// mc/src/store/modules/see/state-types.ts
export enum StateTypes {
    VIDEOREF = "videoRef",
    VIDEO = "video",
}

export enum StateTypes_VIDEO {
    STATE = "state",
    SONG = "song",
    LIST = "list",
    INDEX = "index",
    ORDER = "order",
    VOLUME = "volume",
    FOLD = "fold",
    CURRENTTIME = "currentTime",
    DURATION = "duration",
    LYRIC = "lyric",
}